requirements.txt
streamlit==1.32.0
pandas==2.2.0
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
spacy==3.7.2
wordcloud==1.8.1
openai==1.12.0
groq==0.4.2
streamlit-extras==0.3.6
python-dotenv==1.0.0
matplotlib
numpy
